backend/pytest/requirements.txt
pytest
requests
python-dotenv
faker